pip_requirements.txt
wakepy
django-nested-inline
django-bootstrap-v5
django-crispy-forms
crispy-bootstrap5
ibm-db-django